requirements.txt
packaging>=19.0
setuptools>=41.0.0
wheel
requests>=2.22.0
gitpython>=3.0.0
twine>=3.0.0
nrn-glia>=0.1.8
appdirs>=1.4.3
